[scheduler][feat] Generic task scheduler
Scheduler creates a pool of threads. It then takes in a function and argument list. It then submits the task to one of the threads by adding it to a job queue. The thread when idle, picks up one of the job and executes it. The result is communicated vai futures.

Test: Scheduled few trivial functions from main